refactor: change the initial JPQL to a native query

+++ part-2/src/main/java/link/yauritux/xphrtht/core/domain/dto/EmployeeTimeTrackingReportDto.java
@@ -1,5 +1,7 @@
 package link.yauritux.xphrtht.core.domain.dto;
 
+import java.math.BigDecimal;
+
 /**
  * @author Yauri Attamimi
  * @version 0.1
@@ -7,5 +9,5 @@ package link.yauritux.xphrtht.core.domain.dto;
  * @param projectName the project name
  * @param totalHours total hours spent by the employee on the project
  */
-public record EmployeeTimeTrackingReportDto(String employeeName, String projectName, Double totalHours) {
+public record EmployeeTimeTrackingReportDto(String employeeName, String projectName, BigDecimal totalHours) {
 }
